requirements_whisperx_cuda.txt
# WhisperX CUDA Environment Requirements
# Created: October 29, 2025
# Python: 3.10
# Installation: pip install -r requirements_whisperx_cuda.txt
# Note: PyTorch should be installed via conda first (see CONDA_WHISPERX_SETUP.md)

# ========================================
# Core Deep Learning (Install via Conda)
# ========================================
# conda install pytorch::pytorch pytorch::torchvision pytorch::torchaudio pytorch-cuda=12.1 -c pytorch -c nvidia -y
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

# ========================================
# Scientific Computing (Install via Conda)
# ========================================
# conda install numpy=1.26.4 scipy scikit-learn opencv ffmpeg -y
numpy==1.26.4
scipy==1.14.1
scikit-learn==1.6.1
opencv-python==4.9.0.80
matplotlib==3.10.1
pandas==2.2.3

# ========================================
# Speech Processing & WhisperX
# ========================================
whisperx==3.7.4
faster-whisper==1.2.0
pyannote.audio==3.1.1
librosa==0.10.2.post1
soundfile==0.12.1
pydub==0.25.1

# ========================================
# Korean Language Processing
# ========================================
g2pk==0.9.4
konlpy==0.6.0
jamo==0.4.1

# ========================================
# Transformers & NLP
# ========================================
transformers==4.48.1
huggingface-hub==0.36.0
tokenizers==0.20.3

# ========================================
# Audio/Video Processing
# ========================================
av==12.3.0
ffmpeg-python==0.2.1
Pillow==11.1.0

# ========================================
# Deep Learning Utilities
# ========================================
pytorch-lightning==2.5.5
torchmetrics==1.8.2
pytorch-metric-learning==2.9.0
ctranslate2==4.6.0

# ========================================
# Development & Testing
# ========================================
jupyter==1.1.1
jupyterlab==4.4.6
pytest==8.4.1
tensorboard==2.18.0

# ========================================
# Utilities
# ========================================
pydantic==2.10.6
pyyaml==6.0.2
python-dotenv==1.0.1
tqdm==4.67.1
requests==2.32.3
